Improved quality

+++ lib/Utils/Configuration.hpp
@@ -14,8 +14,8 @@
 const uint8_t MICRO_STEP = 2;
 
 constexpr uint16_t MAX_SPEED = 1750 * MICRO_STEP;
-constexpr uint16_t BALL_ACCELERATION = 20000 * MICRO_STEP;
-constexpr uint16_t PADDLE_ACCELERATION = 20000 * MICRO_STEP;
+constexpr uint32_t BALL_ACCELERATION = 20000 * MICRO_STEP;
+constexpr uint32_t PADDLE_ACCELERATION = 20000 * MICRO_STEP;
 constexpr uint16_t PADDLE_MAX_SPEED = 1400 * MICRO_STEP;
 constexpr uint16_t CALIBRATION_SPEED = 300 * MICRO_STEP;
 /// @brief returns PI/180
